chore(add example error to exercise-01-numbers)

+++ exercises/exercise-01-numbers/start.ts
@@ -1,11 +1,11 @@
 import { UInt32, UInt64 } from "o1js";
 
 // Defining two UInt32 numbers
-const a: UInt64 = UInt64.from(10)
+const a: UInt32 = UInt32.from(10)
 const b: UInt64 = UInt64.from(5_000_000_000);
 
 // Trying to add one UInt64 number to a UInt32 number (incorrect type assignment)
-const result: UInt64 = a.add(b); // This will cause a type error
+const result: UInt32 = a.add(b); // This will cause a type error
 
 console.log(`The result of adding a and b is: ${result} 🎉`);
